EQU Listの出力を修正
