+++ AILZ80ASM/Package.cs
@@ -10,7 +10,7 @@ using System.Text;
 
 namespace AILZ80ASM
 {
-    public class Package
+    public partial class Package
     {
         private List<FileItem> FileItems { get; set; } = new List<FileItem>();
         private AsmOption AssembleOption { get; set; } = default;
@@ -376,285 +376,6 @@ namespace AILZ80ASM
                 default:
                     throw new NotImplementedException($"指定の出力形式は選択できません。{outputFile.Key}");
             }
-        }
-
-        public bool DiffOutput(Dictionary<AsmEnum.FileTypeEnum, FileInfo> outputFiles)
-        {
-            var result = true;
-            foreach (var item in outputFiles)
-            {
-                if (!item.Value.Exists)
-                {
-                    Trace.WriteLine($"{item.Value.Name}: 不一致 ファイルが見つかりません。");
-                    continue;
-                }
-
-                using var fileStream = item.Value.OpenRead();
-
-                try
-                {
-                    result &= DiffOutput(fileStream, item);
-                }
-                catch (Exception ex)
-                {
-                    result = false;
-                    Trace.TraceError(ex.Message);
-                }
-                fileStream.Close();
-            }
-            Trace.WriteLine("");
-            return result;
-        }
-
-        public bool DiffOutput(Stream stream, KeyValuePair<AsmEnum.FileTypeEnum, FileInfo> outputFile)
-        {
-            using var assembledStream = new MemoryStream();
-            using var originalStream = new MemoryStream();
-            stream.CopyTo(originalStream);
-
-            SaveOutput(assembledStream, outputFile);
-            return DiffOutput(originalStream.ToArray(), assembledStream.ToArray(), outputFile);
-        }
-
-        public bool DiffOutput(byte[] original, byte[] assembled, KeyValuePair<AsmEnum.FileTypeEnum, FileInfo> outputFile)
-        {
-            var result = true;
-            var resultString = "一致";
-
-            Trace.Write($"{outputFile.Value.Name}: ");
-
-            if (outputFile.Key == AsmEnum.FileTypeEnum.LST)
-            {
-                // テキスト比較
-                var originals = AILight.AIEncode.GetString(original).Replace("\r", "").Split('\n');
-                var assembleds = AILight.AIEncode.GetString(assembled).Replace("\r", "").Split('\n');
-                if (originals.Length != assembleds.Length)
-                {
-                    resultString = $"不一致 {originals.Length:0} -> {assembleds.Length:0} 行数";
-                    resultString += $"{Environment.NewLine}";
-                    result = false;
-                }
-                else
-                {
-                    var byteDiffCounter = 0;
-                    var tmpResultStream = "";
-
-                    for (var index = 0; index < originals.Length && index < assembleds.Length; index++)
-                    {
-                        if (originals[index] != assembleds[index])
-                        {
-                            if (byteDiffCounter < 100)
-                            {
-                                tmpResultStream += $"{Environment.NewLine}{index + 1:#0}: {originals[index]}";
-                                tmpResultStream += $"{Environment.NewLine}{index + 1:#0}: {assembleds[index]}";
-                            }
-                            byteDiffCounter++;
-                        }
-                    }
-                    if (byteDiffCounter > 0)
-                    {
-                        resultString = $"不一致 ( {byteDiffCounter:0}件 / 全体:{originals.Length:0}行 )" + tmpResultStream;
-                        resultString += $"{Environment.NewLine}";
-                        result = false;
-                    }
-                }
-
-
-                AILight.AIEncode.IsSHIFT_JIS(assembled);
-
-            }
-            else
-            {
-                // バイナリー比較
-                if (original.Length != assembled.Length)
-                {
-                    resultString = $"不一致 {original.Length:0} -> {assembled.Length:0} bytes";
-                    resultString += $"{Environment.NewLine}";
-                    result = false;
-                }
-                else
-                {
-                    var byteDiffCounter = 0;
-                    var tmpResultStream = "";
-                    for (var index = 0; index < original.Length; index++)
-                    {
-                        if (original[index] != assembled[index])
-                        {
-                            if (byteDiffCounter < 100)
-                            {
-                                tmpResultStream += $"{Environment.NewLine}{index:X6}: {original[index]:X2} -> {assembled[index]:X2}";
-                            }
-                            byteDiffCounter++;
-                        }
-                    }
-                    if (byteDiffCounter > 0)
-                    {
-                        resultString = $"不一致 ( {byteDiffCounter:0}件 / 全体:{original.Length:0} bytes )" + tmpResultStream;
-                        resultString += $"{Environment.NewLine}";
-                        result = false;
-                    }
-                }
-            }
-
-            Trace.WriteLine(resultString);
-            return result;
-        }
-
-        public void SaveBin(Stream stream)
-        {
-            var outputAddress = default(UInt32);
-            var binResults = FileItems.SelectMany(m => m.BinResults).OrderBy(m => m.Address.Output);
-
-            // 余白を調整して出力をする
-            foreach (var item in binResults)
-            {
-                if ((item.Address.Output ?? 0) < outputAddress)
-                {
-                    throw new Exception("出力先アドレスが重複したため、BINファイルの出力に失敗ました");
-                }
-                else if ((item.Address.Output ?? 0) != outputAddress)
-                {
-                    // 余白調整
-                    this.AssembleLoad.OutputORGSpace(ref outputAddress, item.Address.Output ?? 0, stream);
-                }
-
-                // 通常出力
-                if (item.Data.Length > 0)
-                {
-                    stream.Write(item.Data, 0, item.Data.Length);
-                    outputAddress += (UInt32)item.Data.Length;
-                }
-            }
-        }
-
-        public void SaveT88(Stream stream, string outputFilename)
-        {
-            using var memoryStream = new MemoryStream();
-            SaveBin(memoryStream);
-
-            var address = default(UInt16);
-            if (AssembleLoad.Share.AsmORGs.Count >= 2)
-            {
-                address = AssembleLoad.Share.AsmORGs.Skip(1).First().ProgramAddress;
-            }
-
-            var binaryWriter = new IO.T88BinaryWriter(outputFilename, address, memoryStream.ToArray(), stream);
-            binaryWriter.Write();
-        }
-
-        public void SaveCMT(Stream stream)
-        {
-            using var memoryStream = new MemoryStream();
-            SaveBin(memoryStream);
-
-            var address = default(UInt16);
-            if (AssembleLoad.Share.AsmORGs.Count >= 2)
-            {
-                address = AssembleLoad.Share.AsmORGs.Skip(1).First().ProgramAddress;
-            }
-
-            var binaryWriter = new IO.CMTBinaryWriter(address, memoryStream.ToArray(), stream);
-            binaryWriter.Write();
-        }
-
-        public void SaveSYM(FileInfo symbol)
-        {
-            using var fileStream = symbol.OpenWrite();
-
-            SaveSYM(fileStream);
-
-            fileStream.Close();
-        }
-
-        public void SaveSYM(Stream stream)
-        {
-            using var memoryStream = new MemoryStream();
-            using var streamWriter = new StreamWriter(memoryStream, AsmLoad.GetEncoding(AssembleLoad.AssembleOption.DecidedOutputEncodeMode));
-
-            var title = $";{ProductInfo.ProductLongName}, SYM";
-            streamWriter.WriteLine(title);
-
-            AssembleLoad.OutputLabels(streamWriter);
-
-            streamWriter.Flush();
-            memoryStream.Position = 0;
-            memoryStream.CopyTo(stream);
-        }
-
-        public void SaveEQU(FileInfo equal)
-        {
-            using var fileStream = equal.OpenWrite();
-
-            SaveEQU(fileStream);
-
-            fileStream.Close();
-        }
-
-        public void SaveEQU(Stream stream)
-        {
-            using var memoryStream = new MemoryStream();
-            using var streamWriter = new StreamWriter(memoryStream, AsmLoad.GetEncoding(AssembleLoad.AssembleOption.DecidedOutputEncodeMode));
-
-            var title = $";{ProductInfo.ProductLongName}, EQU";
-            streamWriter.WriteLine(title);
-
-            AssembleLoad.OutputEqualLabels(streamWriter);
-
-            streamWriter.Flush();
-            memoryStream.Position = 0;
-            memoryStream.CopyTo(stream);
-        }
-
-        public void SaveADR(FileInfo equal)
-        {
-            using var fileStream = equal.OpenWrite();
-
-            SaveADR(fileStream);
-
-            fileStream.Close();
-        }
-
-        public void SaveADR(Stream stream)
-        {
-            using var memoryStream = new MemoryStream();
-            using var streamWriter = new StreamWriter(memoryStream, AsmLoad.GetEncoding(AssembleLoad.AssembleOption.DecidedOutputEncodeMode));
-
-            var title = $";{ProductInfo.ProductLongName}, ADR";
-            streamWriter.WriteLine(title);
-
-            AssembleLoad.OutputAddrLabels(streamWriter);
-
-            streamWriter.Flush();
-            memoryStream.Position = 0;
-            memoryStream.CopyTo(stream);
-        }
-
-        public void SaveLST(FileInfo list)
-        {
-            using var fileStream = list.OpenWrite();
-
-            SaveLST(fileStream);
-
-            fileStream.Close();
-        }
-
-        public void SaveLST(Stream stream)
-        {
-            using var memoryStream = new MemoryStream();
-            using var streamWriter = new StreamWriter(memoryStream, AsmLoad.GetEncoding(AssembleLoad.AssembleOption.DecidedOutputEncodeMode));
-
-            AssembleLoad.ListedFileClear();
-            var lineIndex = 2;
-            var title = $";{ProductInfo.ProductLongName}, LST:{AssembleLoad.AssembleOption.ListMode}:{AssembleLoad.AssembleOption.TabSize}";
-            streamWriter.WriteLine(AsmList.CreateSource(title).ToString(AssembleLoad.AssembleOption.ListMode, AssembleLoad.AssembleOption.TabSize));
-
-            foreach (var item in FileItems)
-            {
-                item.SaveList(streamWriter, ref lineIndex);
-            }
-            streamWriter.Flush();
-            memoryStream.Position = 0;
-            memoryStream.CopyTo(stream);
         }
 
         /// <summary>
